refactor: extract pkg/kube from instance kubernetes.go
The Kubernetes client and its operations move out of pkg/instance into a new leaf package pkg/kube so stack definitions can reference operation implementations without an import cycle (pkg/instance imports pkg/stack). This is a behavior-preserving move: kubernetesService becomes kube.Client with exported fields, its methods are exported, restart is split into RestartDeployment/RestartStatefulSet with the KubernetesResource switch relocated into instance.Service.Restart, and DecryptYaml plus the cluster resource and ingress/cert discovery helpers move along. pkg/group now depends on pkg/kube for FindResources instead of pkg/instance. A temporary NewKubernetesService bridge in pkg/instance keeps the database pod-exec wiring and filestore callers compiling; it is removed in the follow-up component operations change.
