+++ pkg/instance/filestore_test.go
@@ -10,6 +10,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/dhis2-sre/im-manager/pkg/kube"
 	"github.com/dhis2-sre/im-manager/pkg/model"
 	"github.com/minio/minio-go/v7"
 	"github.com/stretchr/testify/assert"
@@ -291,17 +292,17 @@ func TestGetPodByLabels(t *testing.T) {
 		Namespace: "grp",
 		Labels:    map[string]string{"im-type": "minio", "im-deployment-id": "9"},
 	}}
-	ks := kubernetesService{client: fake.NewSimpleClientset(minioPod, otherPod)}
+	ks := kube.Client{Clientset: fake.NewSimpleClientset(minioPod, otherPod)}
 
-	pod, err := ks.getPodByLabels(map[string]string{"im-type": "minio", "im-deployment-id": "7"})
+	pod, err := ks.GetPodByLabels(map[string]string{"im-type": "minio", "im-deployment-id": "7"})
 	require.NoError(t, err)
 	assert.Equal(t, "test-1-minio-abc", pod.Name)
 	assert.Equal(t, "grp", pod.Namespace)
 }
 
 func TestGetPodByLabelsNotFound(t *testing.T) {
-	ks := kubernetesService{client: fake.NewSimpleClientset()}
-	_, err := ks.getPodByLabels(map[string]string{"im-type": "minio", "im-deployment-id": "7"})
+	ks := kube.Client{Clientset: fake.NewSimpleClientset()}
+	_, err := ks.GetPodByLabels(map[string]string{"im-type": "minio", "im-deployment-id": "7"})
 	require.Error(t, err)
 }
